requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
requests==2.32.3
urllib3==2.4.0
